added the ability to schedule jobs and added a dummy job to fill the storage
