+++ GUI/UI/frontend/tasks.py
@@ -1,0 +1,4 @@
+def scheduled_job():
+    f = open('/app/demo.txt','a')
+    f.write("i'm running\n")
+    f.close()
