+++ GUI/requirements.txt
@@ -6,6 +6,7 @@ cffi==1.16.0
 Django==5.0.3
 django-bootstrap-datepicker-plus==5.0.5
 django-bootstrap5==24.2
+django-crontab==0.7.1
 django-filter==24.2
 django-tables2==2.7.0
 gunicorn==22.0.0
